Improve correctness assertion in test_hca_indexer (fixes #125)
Sort lists returned in transformers for determinism
Refactor freeze/thaw out of subscribe.py
Add canned expected output files
